finaly split writing header and payload part
